requirements.txt
faiss-cpu
sentence-transformers
PyPDF2
fastapi
uvicorn
requests
